+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Checkbox/CheckboxPicker/CheckboxPickerConfiguration.swift
@@ -145,7 +145,7 @@ struct CheckboxPickerConfiguration: View {
             OUDSSwitchItem("app_components_common_error_label", isOn: $configurationModel.isError)
                 .disabled(!configurationModel.isEnabled || configurationModel.isReadOnly)
 
-            OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $configurationModel.isReadOnly)
+            OUDSSwitchItem("app_components_common_readOnly_label", isOn: $configurationModel.isReadOnly)
                 .disabled(!configurationModel.isEnabled || configurationModel.isError)
 
             OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $configurationModel.hasDivider)