+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Chip/Filter/FilterChipPage.swift
@@ -48,7 +48,7 @@ struct FilterChipDemo: View {
             switch configurationModel.layout {
             case .iconOnly:
                 OUDSFilterChip(icon: Image(decorative: "ic_heart"),
-                               accessibilityLabel: "app_components_button_icon_a11y".localized(),
+                               accessibilityLabel: "app_components_common_icon_a11y".localized(),
                                selected: configurationModel.selected)
                 {
                     configurationModel.selected.toggle()