+++ DesignToolbox/DesignToolbox/Pages/Components/Controls/Chip/Suggestion/SuggestionChipPage.swift
@@ -49,7 +49,7 @@ struct SuggestionChipDemo: View {
             switch configurationModel.layout {
             case .iconOnly:
                 OUDSSuggestionChip(icon: Image(decorative: "ic_heart"),
-                                   accessibilityLabel: "app_components_button_icon_a11y".localized()) {}
+                                   accessibilityLabel: "app_components_common_icon_a11y".localized()) {}
             case .textOnly:
                 OUDSSuggestionChip(text: configurationModel.text) {}
             case .textAndIcon: